fix mediaserver crash in ACodec::getPortFormat
If the sum of rect.nLeft and rect.nWidth equals to 0, overflow will
happen by minus 1, for unsigned-integer-overflow and
signed-integer-overflow flag was set when compile in Android N.
If we cast the sum to int32_t before minus 1, overflow will not happen.
We should not return err in this case, for omx decoder may report a
message of OMX_EventPortSettingsChanged to inform change of port setting.

Test: NA

Change-Id: I94fdc457233f6424001fa1883acaedc26ccf33b4
Signed-off-by: zengjing <zengjing@xiaomi.com>
Signed-off-by: mydongistiny <jaysonedson@gmail.com>