CameraService: fix torch mode status logic
The status of the torch mode of camera HAL and the framework may be
inconsistant. It results from that the framework forces to update
the status without receiving any notification from HAL, even if camera
HAL supports the callback of the torch mode status.

Fix the logic as below:
1. HAL Version >= CAMERA_MODULE_API_VERSION_2_4
a. A previously-disconnected camera device becomes connected, or
 a previously-connected camera becomes disconnected:
   In these cases, the framework will assume the default torch mode
   status according to the spec. Camera HAL does not need to invoke
   any torch mode callback.
   Thus, keep the original updates from the following path:
    * CameraService::onDeviceStatusChanged() -> updateStatus()

b. Open or close a camera device:
   Camera HAL is responsible for updating the torch mode status by
   invoking a torch mode callback. The framework should not update
   the torch mode status without any notification from HAL.
   Thus, the updates of torch mode status from the following path
   must be disabled:
    * CameraService::BasicClient::startCameraOps() -> updateStatus()
    * CameraService::BasicClient::finishCameraOps() -> updateStatus()

2. HAL Version < CAMERA_MODULE_API_VERSION_2_4
Camera framework is responsible for updating the torch mode status
since camera HAL has no way to notify the framework.

Test: manual testing
Signed-off-by: Jonas Lai <jonas.lai@mediatek.com>
Change-Id: Iee0d10371a41f6609139efa54336e2507496ec53
Signed-off-by: mydongistiny <jaysonedson@gmail.com>